Separated BuildTargetServerIntegrationTest into two separate tests.
Changes:
- Removed BuildTargetServerIntegrationTest
- Added IntegrationTest. Base class for integration tests containing the setup.
- Added BuildTargetServiceIntegrationTest. Subclass of IntegrationTest responsible for handling build target integration tests.
- Added LifecycleServiceIntegrationTest. Subclass of IntegrationTest responsible for handling lifecycle integration tests.
